src/test/java/org/spargonaut/maxConnectFour/referees/InteractiveRefereeTest.java
package org.spargonaut.maxConnectFour.referees;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;
import org.spargonaut.maxConnectFour.gameboard.GameBoard;

import java.io.ByteArrayOutputStream;
import java.io.PrintStream;
import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;

import static org.hamcrest.CoreMatchers.startsWith;
import static org.junit.Assert.assertThat;

public class InteractiveRefereeTest {

    private final ByteArrayOutputStream outContent = new ByteArrayOutputStream();

    @Before
    public void setUp() {
        System.setOut(new PrintStream(outContent));
    }

    @After
    public void tearDown() {
        System.setOut(null);
    }

    @Test
    public void shouldPrintOutTheIntialGameBoardStateAtTheBeginningOfTheGame() {

        GameBoard gameBoard = createEmptyGameBoard();
        InteractiveReferee interactiveReferee = new InteractiveReferee(gameBoard);

        String expectedOutput = "--------------------------------------------------------------------------------\n" +
                "\n" +
                "Max Connect Four Client - INTERACTIVE Mode\n" +
                "\n" +
                "   1 2 3 4 5 6 7   <---  Column numbers\n" +
                " -----------------\n" +
                "_|               |_\n" +
                "_|               |_\n" +
                "_|               |_\n" +
                "_|               |_\n" +
                "_|               |_\n" +
                "_|               |_\n" +
                " -----------------\n" +
                "   1 2 3 4 5 6 7   <---Column numbers\n" +
                "Scores:\n" +
                " Player1: 0\n" +
                " Player2: 0";

        interactiveReferee.play();

        assertThat(outContent.toString(), startsWith(expectedOutput));

    }

    private GameBoard createEmptyGameBoard() {
        List<List<Integer>> playboard = new ArrayList<>();
        playboard.add(createBlankRow());
        playboard.add(createBlankRow());
        playboard.add(createBlankRow());
        playboard.add(createBlankRow());
        playboard.add(createBlankRow());
        playboard.add(createBlankRow());
        return new GameBoard(playboard);
    }

    private List<Integer> createBlankRow() {
        return Arrays.asList(0, 0, 0, 0, 0, 0, 0);
    }
}